Feature/migrate notebook to python modules (#2)
## ✨ Requirement

Migrate the exploration.ipynb notebook as a python module that can be
use later in an API to generate predictions

## ☑️ Change Type

- [x] ✨ New Feature

## 🛠 What's implemented on this PR

The initial scaffolding on model.py is completed based on the
development found on exploration.ipynb. And is used on train_pipeline.py
to generate a training pipeline that saves on disk a trained model and
it's corresponding metrics (over a test set from the original data). The
pipeline versions the models and metrics based on date and hour of
execution, and the model is intended to be loaded in the API to generate
embeddings.

---------

Co-authored-by: Diego Alvarez <diego.alvarezmonroy@mercadolibre.com.co>