Feature/predict api (#5)
## ✨ Requirement

Make the trained model available for predictions by an API, using
FastAPI

## ☑️ Change Type

- [x] ✨ New Feature

## 🛠 What's implemented on this PR

An API with FastAPI to deploy the model and make predictions, it applies
validations of input data with Pydantic, loads the trained model from
disk and generates the predictions to be returned in the defined format
(a dictionary). Also in the `Makefile` the url of the deployed API on
GCP is specified.

### Evidence
<img width="858" alt="image"
src="https://github.com/user-attachments/assets/4abe0d6f-2e03-4581-8666-d735a9ae5ddd"
/>

---------

Co-authored-by: Diego Alvarez <diego.alvarezmonroy@mercadolibre.com.co>